Here is the first iteration of the camera, it is the basic camera.  It just allows you to move the camera around by itself by brute force.
For those of you who I haven't talked to in a while, I am in the processes
of commenting the code and switching the matrix functions over to Kevins
matrix class that he developed.  If you have any questions about the code
let me know and I'll do my best to help you out.


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@652 84b32ba4-53c3-423c-be69-77cca6335494