Bug fixed:	VC++ badness and getChildren
This change "should" fix the problems with VC++ not being able to deal with overloaded member functions where one of the functions is templated.  Supposedly if the template function
is listed last VC++ can figure out what to do.

If this does not work then we can go back to the separate function names (yuck!!!).


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@932 84b32ba4-53c3-423c-be69-77cca6335494