Buttons now splice their text (by calling the new FontRenderer::draw) so that it fits inside the button instead of running out along the side.  If the spliced text is too long to fit on multiple lines, well, it spills under the button :\
The FontRenderer has a second version of draw that takes a phui::Size object which is the constraint of the object to draw text to.  This draw function will splcie the text value in order to meet the width constraints of the Size object.

To see the results, look at the event example.


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@1828 84b32ba4-53c3-423c-be69-77cca6335494