We now have an InputSymbol Library that handles SDLKeys and converts them into InputKeys.  If the function that does this looks surprisingly similar to Chad Austin's gluttophuikey function, don't be alarmed (thanks Chad).
We also have an InputParser that parses a cfg file for action to key bindings.  The parser will then take action pointers and their names and bind them to keys read from the cfg file.  The format for the cfg file is KEY=ACTION where KEY is a string handled in the function InputSymbol::StringtoISym and ACTION is a string you have chosen to identify that action as.  For an example, look at GameState.cpp

Finally, I commented out some of the ai stuff to fix the build.  I have the latest and greatest version of loom, and the code still wouldn't compile :(


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@2031 84b32ba4-53c3-423c-be69-77cca6335494