Bug fixed:	Actually use the body's path in when testing against the 		spatial index rather than its bounds at its initial pos. 		This fixes problems where objects were allowed to pass 		through walls. Bug fixed:	Stop sticking! We now "push back" objects a bit from the 		point of collision so that we don't leave them in a 		colliding state for the next frame. Bug fixed:	All objects no longer collide against everybody else during 		the first frame. We now force OpenSG to update the bounding 		volume each time an entity is changed.
Sorry for the tons of debugging output, but I wanted to make this huge
commit now while it works. Big changes here.

Rigid bodies now fire events to listeners when they are changed. This makes
it so that we can update ourscene graph (and thus the body's bounds)
immediately rather than the next frame.

I reverted CollisionDetector back to only reporting the first collision. I
don't know what I was thinking earlier.

Our hack of a ground plane collision is now made before the body change is
"committed" so that the bounding volume is correct for the collision
detector.


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@2167 84b32ba4-53c3-423c-be69-77cca6335494