DeviceHandle is funny.  it assumes that you don't want to configure your device after you create it.
Basically I found a problem where if you configure joystick (starts with
0 inputs), through DeviceHandle->getDevice()->, then the GameInput will
not call the refreshBindings function (and can't because it would be hard
to notify it with the current architecture).

to fix, I added another constructor to take a preconfigured device
pointer (basically allow someone else to create and set up the device).
and I added a refresh() function to remove/re-add the device (forcing
GameInput to refresh bindings).

a little hacky, but it works.  it still requires that the user be
diligent paying attention that they need to refresh the devicehandle...
:(  not something we want to require, but luckily the only client of this
code is the GameKernel drivers so we don't have to worry too much about
it for the users.


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@2414 84b32ba4-53c3-423c-be69-77cca6335494