New feature: Added Visual Studio Project Files. They are broken so please help chad!!! Refactor: Bug fixed: SF Bug#:
Submitted by:
Reviewed by:
Approved by:
Obtained from:
CVS: ----------------------------------------------------------------------
CVS: Please format log message lines to be no longer than the above line.
CVS: ----------------------------------------------------------------------
CVS: Log header strings:
CVS: New feature:     Fill this in if a new feature is being added.
CVS: Refactor:        Fill this in if the change is a refactoring
CVS:                  of the old code.
CVS: Bug fixed:	      Fill this in if a bug is fixed that does not have a
CVS: 		      SourceForge Bug ID.
CVS: SF Bug#:	      Fill this in if a SourceForge Bug ID is affected by
CVS: 		      the change.
CVS:
CVS: Log footer strings:
CVS: Submitted by:    Fill this in if someone else sent in the patch.
CVS: Reviewed by:     Fill this in if someone else reviewed your modification.
CVS: Approved by:     Fill this in if you needed approval for this commit.
CVS: Obtained from:   Fill this in if the change is from third party software.
CVS:


git-svn-id: file:///desktop/SourceForge%20Backup/isugamedev/svnrepos/trunk@2425 84b32ba4-53c3-423c-be69-77cca6335494